# docs/specification/openapi.yaml
openapi: 3.1.0
info:
  title: TestModeller API
  version: 0.1.0
  description: >
    Contract between the Angular UI and the Rust backend. Changes require explicit approval (see AGENTS.md).
    Hierarchy: Project > Component > Feature > Model. A Feature has a scenario description and one or more Models
    (state machines). Test cases belong to a Feature and are assigned to Model states or transitions (steps).
servers:
  - url: /api/v1
...
paths:
...
  /features/{featureId}/test-cases:
    parameters: [ { $ref: '#/components/parameters/FeatureId' } ]
    get:
      tags: [TestCases]
      summary: List test cases of a feature
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
        - { name: modelId, in: query, description: Only test cases assigned to this model, schema: { $ref: '#/components/schemas/Id' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/TestCaseStatus' } }
        - { name: origin, in: query, schema: { $ref: '#/components/schemas/Origin' } }
        - { name: unassigned, in: query, description: Only test cases without any assignment, schema: { type: boolean } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TestCasePage' } } } }
    post:
      tags: [TestCases]
      summary: Create test case manually
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TestCaseInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/TestCase' } } } }
  /test-cases/{testCaseId}:
    parameters: [ { $ref: '#/components/parameters/TestCaseId' } ]
    get:
      tags: [TestCases]
      summary: Get test case with steps and assignments
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TestCase' } } } }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      tags: [TestCases]
      summary: Replace test case including steps
      parameters: [ { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TestCaseInput' } } } }
      responses:
        '200': { description: Saved, content: { application/json: { schema: { $ref: '#/components/schemas/TestCase' } } } }
        '412': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [TestCases]
      summary: Delete test case
      responses:
        '204': { description: Deleted }
  /test-cases/{testCaseId}/move:
    parameters: [ { $ref: '#/components/parameters/TestCaseId' } ]
    post:
      tags: [TestCases]
      summary: Move test case to another feature (assignments are cleared)
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [featureId], properties: { featureId: { $ref: '#/components/schemas/Id' } } } } }
      responses:
        '200': { description: Moved, content: { application/json: { schema: { $ref: '#/components/schemas/TestCase' } } } }
  /test-cases/{testCaseId}/assignments:
    parameters: [ { $ref: '#/components/parameters/TestCaseId' } ]
    get:
      tags: [Assignments]
      summary: List where this test case is assigned
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Assignment' } } } } }
    put:
      tags: [Assignments]
      summary: Replace all assignments of the test case
      requestBody:
        required: true
        content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AssignmentInput' } } } }
      responses:
        '200': { description: Saved, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Assignment' } } } } }
        '422': { $ref: '#/components/responses/Problem' }
  /states/{stateId}/test-cases:
    parameters: [ { $ref: '#/components/parameters/StateId' } ]
    get:
      tags: [Assignments]
      summary: Test cases assigned to a state
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/TestCase' } } } } }
  /states/{stateId}/test-cases/{testCaseId}:
    parameters: [ { $ref: '#/components/parameters/StateId' }, { $ref: '#/components/parameters/TestCaseId' } ]
    put:
      tags: [Assignments]
      summary: Assign test case to state (idempotent)
      responses:
        '204': { description: Assigned }
        '422': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Assignments]
      summary: Remove assignment
      responses:
        '204': { description: Removed }
  /transitions/{transitionId}/test-cases:
    parameters: [ { $ref: '#/components/parameters/TransitionId' } ]
    get:
      tags: [Assignments]
      summary: Test cases assigned to a transition (model step)
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/TestCase' } } } } }
  /transitions/{transitionId}/test-cases/{testCaseId}:
    parameters: [ { $ref: '#/components/parameters/TransitionId' }, { $ref: '#/components/parameters/TestCaseId' } ]
    put:
      tags: [Assignments]
      summary: Assign test case to transition (idempotent)
      responses:
        '204': { description: Assigned }
        '422': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Assignments]
      summary: Remove assignment
      responses:
        '204': { description: Removed }
  /models/{modelId}/test-cases:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    get:
      tags: [Assignments]
      summary: Test cases assigned anywhere in the model, with their assignments
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/TestCase' } } } } }
  /models/{modelId}/generate:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [Generation]
      summary: Generate test cases from the model. Deterministic for model, criterion and seed.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenerationRequest' } } } }
      responses:
        '200': { description: Generated (saved only if save=true), content: { application/json: { schema: { $ref: '#/components/schemas/GenerationResult' } } } }
        '422': { $ref: '#/components/responses/Problem' }
  /models/{modelId}/coverage:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    get:
      tags: [Coverage]
      summary: Coverage of the model by assigned test cases
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Coverage' } } } }
  /features/{featureId}/coverage:
    parameters: [ { $ref: '#/components/parameters/FeatureId' } ]
    get:
      tags: [Coverage]
      summary: Aggregated coverage of a feature
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Coverage' } } } }
  /components/{componentId}/coverage:
    parameters: [ { $ref: '#/components/parameters/ComponentId' } ]
    get:
      tags: [Coverage]
      summary: Aggregated coverage of a component
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Coverage' } } } }
...
components:
  parameters:
...
    ComponentId: { name: componentId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    FeatureId: { name: featureId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    ModelId: { name: modelId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    StateId: { name: stateId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    TransitionId: { name: transitionId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    TestCaseId: { name: testCaseId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Entity `version` for optimistic concurrency
      schema: { type: integer, minimum: 1 }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
...
  responses:
    Problem:
      description: Error (RFC 7807)
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
  schemas:
    Id: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
    Origin: { type: string, enum: [manual, generated, ai] }
    TestCaseStatus: { type: string, enum: [draft, approved, deprecated] }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
...
    Audit:
      type: object
      required: [id, version, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        version: { type: integer, minimum: 1 }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
...
    TestStepInput:
      type: object
      required: [action, expected]
      properties:
        action: { type: string }
        expected: { type: string }
    TestStep:
      allOf:
        - { $ref: '#/components/schemas/TestStepInput' }
        - type: object
          required: [order]
          properties:
            order: { type: integer, minimum: 1 }
    TestCaseInput:
      type: object
      required: [name, steps]
      properties:
        name: { type: string, minLength: 1, maxLength: 300 }
        description: { type: string }
        preconditions: { type: string }
        priority: { type: string, enum: [low, medium, high] }
        status: { $ref: '#/components/schemas/TestCaseStatus' }
        tags: { type: array, items: { type: string } }
        steps: { type: array, items: { $ref: '#/components/schemas/TestStepInput' } }
        assignments: { type: array, items: { $ref: '#/components/schemas/AssignmentInput' } }
    TestCase:
      allOf:
        - { $ref: '#/components/schemas/Audit' }
        - type: object
          required: [featureId, name, steps, origin, assignments]
          properties:
            featureId: { $ref: '#/components/schemas/Id' }
            name: { type: string }
            description: { type: string }
            preconditions: { type: string }
            priority: { type: string, enum: [low, medium, high] }
            status: { $ref: '#/components/schemas/TestCaseStatus' }
            tags: { type: array, items: { type: string } }
            origin: { $ref: '#/components/schemas/Origin' }
            generatedFromModelId: { $ref: '#/components/schemas/Id' }
            steps: { type: array, items: { $ref: '#/components/schemas/TestStep' } }
            assignments: { type: array, items: { $ref: '#/components/schemas/Assignment' } }
    TestCasePage:
      type: object
      properties: { items: { type: array, items: { $ref: '#/components/schemas/TestCase' } }, nextCursor: { type: [string, 'null'] } }
    AssignmentInput:
      type: object
      required: [modelId]
      description: Exactly one of stateId or transitionId. A test case may be assigned to many states/transitions, also across models of its feature.
      properties:
        modelId: { $ref: '#/components/schemas/Id' }
        stateId: { $ref: '#/components/schemas/Id' }
        transitionId: { $ref: '#/components/schemas/Id' }
        stepOrder: { type: integer, minimum: 1, description: Optional test step this assignment refers to }
    Assignment:
      allOf:
        - { $ref: '#/components/schemas/AssignmentInput' }
        - type: object
          required: [testCaseId]
          properties: { testCaseId: { $ref: '#/components/schemas/Id' } }
    CoverageCriterion: { type: string, enum: [state, transition, transition-pair, bounded-paths] }
    GenerationRequest:
      type: object
      required: [criterion]
      properties:
        criterion: { $ref: '#/components/schemas/CoverageCriterion' }
        seed: { type: integer }
        maxPathLength: { type: integer, minimum: 1, description: Required for bounded-paths }
        save: { type: boolean, default: false, description: Persist generated test cases and assign them to the model }
    GenerationResult:
      type: object
      required: [testCases, coverage]
      properties:
        seed: { type: integer }
        testCases: { type: array, items: { $ref: '#/components/schemas/TestCase' } }
        coverage: { $ref: '#/components/schemas/Coverage' }
        skippedPaths: { type: array, items: { type: object, properties: { transitionIds: { type: array, items: { $ref: '#/components/schemas/Id' } }, reason: { type: string } } } }
    Coverage:
      type: object
      required: [states, transitions]
      properties:
        states: { $ref: '#/components/schemas/CoverageCount' }
        transitions: { $ref: '#/components/schemas/CoverageCount' }
        uncoveredStateIds: { type: array, items: { $ref: '#/components/schemas/Id' } }
        uncoveredTransitionIds: { type: array, items: { $ref: '#/components/schemas/Id' } }
    CoverageCount:
      type: object
      required: [covered, total]
      properties: { covered: { type: integer }, total: { type: integer } }
...